Tests/LibWeb/Text/expected/Fetch/image-accept-header.txt

image/avif,image/webp,image/png,image/svg+xml,image/*;q=0.8,*/*;q=0.5
